+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [Release 1.25.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.25.0)
+
+### Features
+- **datastore:** Added logic to retry on sync failure. (#1414)
+
+[See all changes between 1.24.1 and 1.25.0](https://github.com/aws-amplify/amplify-android/compare/release_v1.24.1...release_v1.25.0)
+
+## [Release 1.24.1](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.24.1)
+
+### Bug Fixes
+- handle null values for predicates (#1435)
+
+[See all changes between 1.24.0 and 1.24.1](https://github.com/aws-amplify/amplify-android/compare/release_v1.24.0...release_v1.24.1)
+
 ## [Release 1.24.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.24.0)
 
 ### Features