+++ core-kotlin/CHANGELOG.md
@@ -1,3 +1,17 @@
+## [Release 0.9.0](https://github.com/aws-amplify/amplify-android/releases/tag/release-kotlin_v0.9.0)
+
+### Features
+- **datastore:** Retry sync (#1414)
+
+[See all changes between 0.8.1 and 0.9.0](https://github.com/aws-amplify/amplify-android/compare/release-kotlin_v0.8.1...release-kotlin_v0.9.0)
+
+## [Release 0.8.1](https://github.com/aws-amplify/amplify-android/releases/tag/release-kotlin_v0.8.1)
+
+### Bug Fixes
+- handle null values for predicates (#1435)
+
+[See all changes between 0.8.0 and 0.8.1](https://github.com/aws-amplify/amplify-android/compare/release-kotlin_v0.8.0...release-kotlin_v0.8.1)
+
 ## [Release 0.8.0](https://github.com/aws-amplify/amplify-android/releases/tag/release-kotlin_v0.8.0)
 
 ### Features